Refactor Wordle component: Simplify input handling and remove useWordleInput hook
- Removed `useWordleInput` custom hook and integrated its logic directly into the Wordle component.
- Refactored `handleInput` to manage key input handling, submission, and input resetting within the component.
- Moved the `ENTER` key submission logic outside of `setInputValue` to prevent duplicate submissions caused by state changes.
- Introduced `useState` to handle `inputValue` directly in the Wordle component.
- Removed unnecessary `resetInput` logic in favor of clearing `inputValue` state directly.
- Cleaned up `wordle-actions.ts` by removing a redundant `console.log` statement.
